+++ lib/features/local_send/widgets/import_store_archive_dialog.dart
@@ -2,8 +2,9 @@ import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:hoplixi/core/app_paths.dart';
 import 'package:hoplixi/core/utils/toastification.dart';
-import 'package:hoplixi/main_db/old/provider/archive_provider.dart';
-import 'package:hoplixi/main_db/old/services/other/archive_service.dart';
+import 'package:hoplixi/main_db/new/providers/archive_provider.dart';
+import 'package:hoplixi/main_db/new/services/archive_service/archive_service.dart';
+
 import 'package:hoplixi/shared/ui/button.dart';
 import 'package:hoplixi/shared/ui/text_field.dart';