+++ lib/features/password_manager/dashboard/screens/notes_graph_screen.dart
@@ -4,7 +4,7 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:go_router/go_router.dart';
 import 'package:hoplixi/core/logger/index.dart';
 import 'package:hoplixi/features/password_manager/dashboard/models/entity_type.dart';
-import 'package:hoplixi/main_db/old/models/graph_data.dart';
+import 'package:hoplixi/main_db/core/models/graph_data.dart';
 import 'package:hoplixi/main_db/new/providers/dao_providers.dart';
 import 'package:hoplixi/routing/paths.dart';