+++ lib/main_db/new/providers/archive_provider.dart
@@ -1,5 +1,5 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:hoplixi/main_db/old/services/other/archive_service.dart';
+import 'package:hoplixi/main_db/new/services/archive_service/archive_service.dart';
 
 /// Провайдер для сервиса архивации хранилищ
 final archiveServiceProvider = Provider<ArchiveService>((ref) {