+++ lib/main_db/old/provider/db_history_provider.dart
@@ -1,8 +1,0 @@
-import 'package:riverpod/riverpod.dart';
-import '../services/db_history_services.dart';
-
-final dbHistoryProvider = FutureProvider<DatabaseHistoryService>((ref) async {
-  final databaseHistoryService = DatabaseHistoryService();
-  await databaseHistoryService.initialize();
-  return databaseHistoryService;
-});