+++ lib/main_db/old/provider/index.dart
@@ -1,8 +1,0 @@
-library;
-
-export 'dao_providers.dart';
-export 'db_history_provider.dart';
-export 'decrypted_files_guard_provider.dart';
-export 'main_store_backup_orchestrator_provider.dart';
-export 'main_store_provider.dart';
-export 'service_providers.dart';